[new-fork][2/N] Json mode failure handling (#523)
This PR allows to fail sequence group with invalid masks. 

It works this way;

1. If there's an invalid mask, we find the relevant sequence group.
2. We find the index and mark relevant seq group as failed
3. llm_engine.py finishes the request and set the "error" field (same as
fork).

Note that the old fork raises an exception for async engine, and not
raising an exception for non-async engine. I currently chose not to
raise exception at all, but I can change it if needed.

This PR is very involving, so we should upstream it asap.